Add error summary examples
